매물등록 메서드 수정, User ID값 Long으로 수정

+++ src/main/java/com/nz/entity/UserEntity.java
@@ -25,7 +25,7 @@ public class UserEntity {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq_member")
-	private Integer memberID;
+	private Long memberID;
 	
 	private String name;
