+++ src/main/java/com/nz/repository/UserRepository.java
@@ -5,10 +5,11 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.nz.entity.UserEntity;
 
-public interface UserRepository extends JpaRepository<UserEntity, Integer>{
+public interface UserRepository extends JpaRepository<UserEntity, Long>{
 	
 	Optional<UserEntity> findByUsername(String username);
 	
 	boolean existsByUsername(String username);
+
 
 }
